requirements/server.requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=server --output-file=requirements/server.requirements.txt --strip-extras
#
aiohttp==3.9.0
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
attrs==23.1.0
    # via aiohttp
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via click
fastapi==0.104.1
    # via weather (setup.py)
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   requests
    #   yarl
joblib==1.3.2
    # via scikit-learn
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.26.2
    # via
    #   pandas
    #   scikit-learn
    #   scipy
    #   weather (setup.py)
pandas==2.1.3
    # via weather (setup.py)
prettytable==3.9.0
    # via weather (setup.py)
pydantic==2.5.2
    # via
    #   fastapi
    #   pydantic-settings
    #   weather (setup.py)
pydantic-core==2.14.5
    # via pydantic
pydantic-settings==2.1.0
    # via weather (setup.py)
pyjwt==2.8.0
    # via twilio
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via
    #   pydantic-settings
    #   weather (setup.py)
pytz==2023.3.post1
    # via pandas
requests==2.31.0
    # via twilio
scikit-learn==1.2.2
    # via weather (setup.py)
scipy==1.11.4
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
threadpoolctl==3.2.0
    # via scikit-learn
timeloop==1.0.2
    # via weather (setup.py)
twilio==8.12.0
    # via weather (setup.py)
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via pandas
urllib3==2.0.7
    # via requests
uvicorn==0.24.0.post1
    # via weather (setup.py)
wcwidth==0.2.12
    # via prettytable
yarl==1.9.3
    # via aiohttp
